pyproject.toml
[project]
name = "local-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-qdrant>=1.1.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.42.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
